+++ request_checker/android/app/src/main/kotlin/com/example/request_checker/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.request_checker
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}